PowerAnalytics

API documentation

PowerAnalytics

Index

Exported

Internal

Index

PowerAnalytics.categorize_data

PowerAnalytics.combine_categories

PowerAnalytics.get_generator_category

PowerAnalytics.get_generator_mapping

PowerAnalytics.make_fuel_dictionary

Exported

PowerAnalytics.categorize_data — Method

Re-categorizes data according to an aggregation dictionary

makes no guarantee of complete data collection *

Example

aggregation = PA.make_fuel_dictionary(results_uc.system)
categorize_data(gen_uc.data, aggregation)

PowerAnalytics.make_fuel_dictionary — Method

generators = make_fuel_dictionary(system::PSY.System, mapping::Dict{NamedTuple, String})

This function makes a dictionary of fuel type and the generators associated.

Arguments

sys::PSY.System: the system that is used to create the results

results::IS.Results: results

Key Words

categories::Dict{String, NamedTuple}: if stacking by a different category is desired

Example

results = solveopmodel!(OpModel) generators = makefueldictionary(sys)

Internal

PowerAnalytics.combine_categories — Method

aggregates and combines data into single DataFrame

Example

PG.combine_categories(gen_uc.data)

PowerAnalytics.get_generator_category — Method

Return the generator category for this fuel and unit_type.

PowerAnalytics.get_generator_mapping — Function

Return a dict where keys are a tuple of input parameters (fuel, unit_type) and values are generator types.