Exported

PowerAnalytics.AGG_META_KEY — Constant

Column metadata key whose value, if any, is additional information to be passed to aggregation functions. Values of nothing are equivalent to absence of the entry.

PowerAnalytics.DATETIME_COL — Constant

Name of the column that represents the time axis in computed DataFrames

PowerAnalytics.META_COL_KEY — Constant

Column metadata key whose value signifies whether the column is metadata. Metadata columns are excluded from get_data_cols and similar and can be used to represent things like a time aggregation.

PowerAnalytics.RESULTS_COL — Constant

Name of a column that represents whole-of-Results data

PowerAnalytics.SYSTEM_COL — Constant

Name of a column that represents whole-of-System data

PowerAnalytics.ComponentSelectorTimedMetric — Type

Time series Metrics defined on ComponentSelectors.

PowerAnalytics.ComponentTimedMetric — Type

ComponentSelectorTimedMetrics implemented by evaluating a function on each Component.

Arguments

name::String: the name of the Metric

eval_fn: a function with signature (::IS.Results, ::Component; start_time::Union{Nothing, DateTime}, len::Union{Int, Nothing}) that returns a DataFrame representing the results for that Component

component_agg_fn: optional, a function to aggregate results between Components/ComponentSelectors, defaults to sum

time_agg_fn: optional, a function to aggregate results across time, defaults to sum

component_meta_agg_fn: optional, a callable to aggregate metadata across components, defaults to sum

time_meta_agg_fn: optional, a callable to aggregate metadata across time, defaults to sum

eval_zero: optional and rarely filled in, specifies what to do in the case where there are no components to contribute to a particular group; defaults to nothing, in which case the data is filled in from the identity element of component_agg_fn

PowerAnalytics.CustomTimedMetric — Type

ComponentSelectorTimedMetrics implemented without drilling down to the base Components, just call the eval_fn directly.

Arguments

name::String: the name of the Metric

eval_fn: a callable with signature (::IS.Results, ::Union{ComponentSelector, Component}; start_time::Union{Nothing, DateTime}, len::Union{Int, Nothing}) that returns a DataFrame representing the results for that Component

time_agg_fn: optional, a callable to aggregate results across time, defaults to sum

time_meta_agg_fn: optional, a callable to aggregate metadata across time, defaults to sum

PowerAnalytics.Metric — Type

The basic type for all Metrics.

PowerAnalytics.ResultsTimelessMetric — Type

Timeless Metrics with a single value per IS.Results instance

Arguments

- `name::String`: the name of the `Metric`
- `eval_fn`: a callable with signature `(::IS.Results,)` that returns a `DataFrame`
  representing the results

PowerAnalytics.SystemTimedMetric — Type

Time series Metrics defined on Systems.

Arguments

name::String: the name of the Metric

eval_fn: a callable with signature (::IS.Results; start_time::Union{Nothing, DateTime}, len::Union{Int, Nothing}) that returns a DataFrame representing the results

time_agg_fn: optional, a callable to aggregate results across time, defaults to sum

time_meta_agg_fn: optional, a callable to aggregate metadata across time, defaults to sum

PowerAnalytics.TimedMetric — Type

Time series Metrics.

PowerAnalytics.TimelessMetric — Type

Scalar-in-time Metrics.

PowerAnalytics.aggregate_time — Method

Given a DataFrame like that produced by compute_all, group by a function of the time axis, apply a reduction, and report the resulting aggregation indexed by the first timestamp in each group.

Arguments

df::DataFrames.AbstractDataFrame: the DataFrame to operate upon

groupby_fn = nothing: a callable that can be passed a DateTime; two rows will be in the same group iff their timestamps produce the same result under groupby_fn. Note that groupby_fn = month puts January 2023 and January 2024 into the same group whereas groupby_fn=(x -> (year(x), month(x))) does not.

groupby_col::Union{Nothing, AbstractString, Symbol} = nothing: specify a column name to report the result of groupby_fn in the output DataFrame, or nothing to not

agg_fn = nothing: by default, the aggregation function (sum/mean/etc.) is specified by the Metric, which is read from the metadata of each column. If this metadata isn't found, one can specify a default aggregation function like sum here; if nothing, an error will be thrown.

PowerAnalytics.categorize_data — Method

Re-categorizes data according to an aggregation dictionary

makes no guarantee of complete data collection *

Example

aggregation = PA.make_fuel_dictionary(results_uc.system)
categorize_data(gen_uc.data, aggregation)

PowerAnalytics.compose_metrics — Function

Given a list of metrics and a function that applies to their results to produce one result, create a new metric that computes the sub-metrics and applies the function to produce its own result.

Arguments

name::String: the name of the new Metric

reduce_fn: a callable that takes one value from each of the input Metrics and returns a single value that will be the result of this Metric. "Value" means a vector (not a DataFrame) in the case of TimedMetrics and a scalar for TimelessMetrics.

metrics: the input Metrics. It is currently not possible to combine TimedMetrics with TimelessMetrics, though it is possible to combine ComponentSelectorTimedMetrics with SystemTimedMetrics.

PowerAnalytics.compute — Method

Compute the given metric on the groups of the given ComponentSelector within the given set of results, returning a DataFrame with a DateTime column and a data column for each group. Exclude components marked as not available.

Arguments

metric::ComponentTimedMetric: the metric to compute

results::IS.Results: the results from which to fetch data

selector::ComponentSelector: the ComponentSelector on whose subselectors to compute the metric

start_time::Union{Nothing, DateTime} = nothing: the time at which the resulting time series should begin

len::Union{Int, Nothing} = nothing: the number of steps in the resulting time series

PowerAnalytics.compute — Method

Compute the given metric on the given component within the given set of results, returning a DataFrame with a DateTime column and a data column labeled with the component's name.

Arguments

metric::ComponentTimedMetric: the metric to compute

results::IS.Results: the results from which to fetch data

comp::Component: the component on which to compute the metric

start_time::Union{Nothing, DateTime} = nothing: the time at which the resulting time series should begin

len::Union{Int, Nothing} = nothing: the number of steps in the resulting time series

PowerAnalytics.compute — Method

Compute the given metric on the given component within the given set of results, returning a DataFrame with a DateTime column and a data column labeled with the component's name. Exclude components marked as not available.

Arguments

metric::CustomTimedMetric: the metric to compute

results::IS.Results: the results from which to fetch data

comp::Component: the component on which to compute the metric

start_time::Union{Nothing, DateTime} = nothing: the time at which the resulting time series should begin

len::Union{Int, Nothing} = nothing: the number of steps in the resulting time series

PowerAnalytics.compute — Method

Convenience method for compute_all; returns compute(metric, results)

PowerAnalytics.compute — Method

Compute the given metric on the given set of results, returning a DataFrame with a single cell. Exclude components marked as not available.

Arguments

metric::ResultsTimelessMetric: the metric to compute

results::IS.Results: the results from which to fetch data

PowerAnalytics.compute — Method

Convenience method for compute_all; returns compute(metric, results; kwargs...)

PowerAnalytics.compute — Method

Compute the given metric on the System associated with the given set of results, returning a DataFrame with a DateTime column and a data column.

Arguments

metric::SystemTimedMetric: the metric to compute

results::IS.Results: the results from which to fetch data

start_time::Union{Nothing, DateTime} = nothing: the time at which the resulting time series should begin

len::Union{Int, Nothing} = nothing: the number of steps in the resulting time series

PowerAnalytics.compute_all — Function

For each (metric, selector, col_name) tuple in zip(metrics, selectors, col_names), call compute and collect the results in a DataFrame with a single DateTime column. All selectors must yield exactly one group.

Arguments

results::IS.Results: the results from which to fetch data

metrics::Vector{<:TimedMetric}: the metrics to compute

selectors: either a scalar or vector of Nothing/Component/ComponentSelector: the selectors on which to compute the metrics, or nothing for system/results metrics; broadcast if scalar

col_names::Union{Nothing, Vector{<:Union{Nothing, AbstractString}}} = nothing: a vector of names for the columns of ouput data. Entries of nothing default to the result of metric_selector_to_string; names = nothing is equivalent to an entire vector of nothing

kwargs...: pass through to each compute call

PowerAnalytics.compute_all — Function

For each (metric, colname) tuple in `zip(metrics, colnames), call [compute`](@ref) and collect the results in a DataFrame.

Arguments

results::IS.Results: the results from which to fetch data

metrics::Vector{<:TimelessMetric}: the metrics to compute

selectors: either a scalar or vector of Nothing/Component/ComponentSelector: the selectors on which to compute the metrics, or nothing for system/results metrics; broadcast if scalar

col_names::Union{Nothing, Vector{<:Union{Nothing, AbstractString}}} = nothing: a vector of names for the columns of ouput data. Entries of nothing default to the result of metric_selector_to_string; names = nothing is equivalent to an entire vector of nothing

kwargs...: pass through to each compute call

PowerAnalytics.compute_all — Method

For each (metric, selector, col_name) tuple in computations, call compute and collect the results in a DataFrame with a single DateTime column. All selectors must yield exactly one group.

Arguments

results::IS.Results: the results from which to fetch data

computations::(Tuple{<:T, Any, Any} where T <: Union{TimedMetric, TimelessMetric})...: a list of the computations to perform, where each element is a (metric, selector, col_name)wheremetricis the metric to compute,selectoris the ComponentSelector on which to compute the metric ornothingif not relevant, andcol_name` is the name for the output column of data or nothing to use the default

kwargs...: pass through to each compute call

PowerAnalytics.create_problem_results_dict — Function

Accept a directory that contains several results subdirectories (that each contain results, problems, etc. sub-subdirectories) and construct a sorted dictionary from String to SimulationProblemResults where the keys are the subdirectory names and the values are loaded results datasets.

Arguments

results_dir::AbstractString: the directory where results subdirectories can be found

problem::String: the name of the problem to load (e.g., "UC", "ED")

scenarios::Union{Vector{AbstractString}, Nothing} = nothing: a list of scenario subdirectories to load, or nothing to load all the subdirectories

kwargs...: keyword arguments to pass through to get_decision_problem_results

PowerAnalytics.get_agg_meta — Method

Get the column's aggregation metadata; return nothing if there is none.

PowerAnalytics.get_agg_meta — Method

Get the single data column's aggregation metadata; error on multiple data columns.

PowerAnalytics.get_data_cols — Method

Select the names of the data columns of the DataFrame, i.e., those that are not DateTime and not metadata.

PowerAnalytics.get_data_df — Method

Select the data columns of the DataFrame as a DataFrame without copying.

PowerAnalytics.get_data_mat — Method

Select the data columns of the DataFrame as a Matrix with copying.

PowerAnalytics.get_data_vec — Method

Select the data column of the DataFrame as a vector without copying, errors if more than one.

PowerAnalytics.get_time_df — Method

Select the DateTime column of the DataFrame as a one-column DataFrame without copying.

PowerAnalytics.get_time_vec — Method

Select the DateTime column of the DataFrame as a Vector without copying.

PowerAnalytics.hcat_timed_dfs — Method

If the time axes match across all the DataFrames, horizontally concatenate them and remove the duplicate time axes. If not, throw an error

PowerAnalytics.is_col_meta — Method

Check whether a column is metadata

PowerAnalytics.make_fuel_dictionary — Method

generators = make_fuel_dictionary(system::PSY.System, mapping::Dict{NamedTuple, String})

This function makes a dictionary of fuel type and the generators associated.

Arguments

sys::PSY.System: the system that is used to create the results

results::IS.Results: results

Key Words

categories::Dict{String, NamedTuple}: if stacking by a different category is desired

Example

results = solveopmodel!(OpModel) generators = makefueldictionary(sys)

PowerAnalytics.metric_selector_to_string — Method

Canonical way to represent a (Metric, ComponentSelector) or (Metric, Component) pair as a string.

PowerAnalytics.parse_generator_categories — Method

Use parse_generator_mapping_file to parse a generator_mapping.yaml file into a dictionary of ComponentSelector, excluding categories in the 'non_generators' list in metadata

PowerAnalytics.parse_generator_mapping_file — Method

Parse a generator_mapping.yaml file into a dictionary of ComponentSelectors and a dictionary of metadata if present

PowerAnalytics.parse_injector_categories — Method

Use parse_generator_mapping_file to parse a generator_mapping.yaml file into a dictionary of all ComponentSelectors

PowerAnalytics.rebuild_metric — Method

Returns a Metric identical to the input metric except with the changes to its fields specified in the keyword arguments.

Examples

Make a variant of calc_active_power that averages across components rather than summing:

using PowerAnalytics.Metrics
calc_active_power_mean = rebuild_metric(calc_active_power; component_agg_fn = mean)
# (now calc_active_power_mean works as a standalone, callable metric)

PowerAnalytics.set_agg_meta! — Method

Set the column's aggregation metadata.

PowerAnalytics.set_agg_meta! — Method

Set the single data column's aggregation metadata; error on multiple data columns.

PowerAnalytics.set_col_meta! — Function

Mark a column as metadata

PowerAnalytics.unweighted_sum — Method

A version of sum that ignores a second argument, for use where aggregation metadata is at play

PowerAnalytics.weighted_mean — Method

Compute the mean of values weighted by the corresponding entries of weights. Arguments may be vectors or vectors of vectors. A weight of 0 cancels out a value of NaN.

PowerAnalytics.Metrics — Module

PowerAnalytics built-in Metrics. Use names to list what is available.

PowerAnalytics.Metrics.calc_active_power — Constant

Calculate the active power of the specified ComponentSelector

PowerAnalytics.Metrics.calc_active_power_forecast — Constant

Fetch the forecast active power of the specified ComponentSelector

PowerAnalytics.Metrics.calc_active_power_in — Constant

Calculate the active power input to the specified (storage) ComponentSelector

PowerAnalytics.Metrics.calc_active_power_out — Constant

Calculate the active power output of the specified (storage) ComponentSelector

PowerAnalytics.Metrics.calc_capacity_factor — Constant

Calculate the capacity factor (actual production/rated production) of the specified ComponentSelector

PowerAnalytics.Metrics.calc_curtailment — Constant

Calculate the ActivePowerForecast minus the ActivePower of the given ComponentSelector

PowerAnalytics.Metrics.calc_curtailment_frac — Constant

Calculate the Curtailment as a fraction of the ActivePowerForecast of the given ComponentSelector

PowerAnalytics.Metrics.calc_discharge_cycles — Constant

Calculate the number of discharge cycles a storage device has gone through in the time period

PowerAnalytics.Metrics.calc_integration — Constant

Calculate the ActivePower of the given ComponentSelector over the sum of the SystemLoadForecast and the SystemLoadFromStorage

PowerAnalytics.Metrics.calc_is_slack_up — Constant

Calculate whether the given time period has system balance slack up of magnitude greater than 0.0001

PowerAnalytics.Metrics.calc_load_forecast — Constant

Fetch the forecast active load of the specified ComponentSelector

PowerAnalytics.Metrics.calc_load_from_storage — Constant

Calculate the ActivePowerIn minus the ActivePowerOut of the specified (storage) ComponentSelector

PowerAnalytics.Metrics.calc_net_load_forecast — Constant

SystemLoadForecast minus ActivePowerForecast of the given ComponentSelector

PowerAnalytics.Metrics.calc_production_cost — Constant

Calculate the active power output of the specified ComponentSelector

PowerAnalytics.Metrics.calc_shutdown_cost — Constant

Calculate the shutdown cost of the specified ComponentSelector

PowerAnalytics.Metrics.calc_startup_cost — Constant

Calculate the startup cost of the specified ComponentSelector

PowerAnalytics.Metrics.calc_stored_energy — Constant

Calculate the energy stored in the specified (storage) ComponentSelector

PowerAnalytics.Metrics.calc_sum_bytes_alloc — Constant

Sum the bytes allocated to the optimization problems

PowerAnalytics.Metrics.calc_sum_objective_value — Constant

Sum the objective values achieved in the optimization problems

PowerAnalytics.Metrics.calc_sum_solve_time — Constant

Sum the solve times taken by the optimization problems

PowerAnalytics.Metrics.calc_system_load_forecast — Constant

Fetch the forecast active load of all the ElectricLoad Components in the system

PowerAnalytics.Metrics.calc_system_load_from_storage — Constant

Fetch the LoadFromStorage of all storage in the system

PowerAnalytics.Metrics.calc_system_slack_up — Constant

Calculate the system balance slack up

PowerAnalytics.Metrics.calc_total_cost — Constant

Calculate the production cost of the specified ComponentSelector, plus the startup and shutdown costs if they are defined

PowerAnalytics.Selectors — Module

PowerAnalytics built-in ComponentSelectors. Use names to list what is available.

PowerAnalytics.Selectors.all_loads — Constant

A ComponentSelector representing all the electric load in a System

PowerAnalytics.Selectors.all_storage — Constant

A ComponentSelector representing all the storage in a System

PowerAnalytics.Selectors.categorized_generators — Constant

A single ComponentSelector representing the generators in a System (no storage or load) grouped by the categories in generator_mapping.yaml

PowerAnalytics.Selectors.categorized_injectors — Constant

A single ComponentSelector representing the static injectors in a System grouped by the categories in generator_mapping.yaml

PowerAnalytics.Selectors.generator_categories — Constant

A dictionary of ComponentSelectors, each of which corresponds to one of the categories in generator_mapping.yaml, only considering the components and categories that represent generators (no storage or load)

PowerAnalytics.Selectors.injector_categories — Constant

A dictionary of ComponentSelectors, each of which corresponds to one of the static injector categories in generator_mapping.yaml